requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pdfplumber>=0.11.0
PyPDF2>=3.0.1
python-docx>=1.1.2
pandas>=2.2.0
openpyxl>=3.1.2
